--- FileManager/FileManager/Program.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace FileManager
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            FileManager fileManager = new FileManager();
-            fileManager.Process();
-        }
-    }
-}
